Base function to get status of last report generated

Function to check the status of the last API request sent with get_raster().

Usage

get_last_report(key = gfw_auth())

Arguments

key

Authorization token. Can be obtained with gfw_auth() function

Details

The get_last_report() function will tell you if the APIs are still processing your request and will download the results if the request has finished successfully. You will receive an error message if the request finished but resulted in an error or if it's been >30 minutes since the last report was generated using get_raster().

For more information, see the https://globalfishingwatch.org/our-apis/documentation#get-last-report-generated.

Examples

if (FALSE) { # \dontrun{
get_last_report(key = gfw_auth())
} # }