Base function to get vessel information from API and convert response to data frame

Base function to get vessel information from API and convert response to data frame

get_vessel_info(
  query = NULL,
  where = NULL,
  ids = NULL,
  includes = c("AUTHORIZATIONS", "OWNERSHIP", "MATCH_CRITERIA"),
  match_fields = NULL,
  registries_info_data = c("ALL"),
  search_type = "search",
  key = gfw_auth(),
  print_request = FALSE,
  ...
)

Arguments

query

When search_type = "search", a length-1 vector with the identity variable of interest, MMSI, IMO, call sign or ship name.

where

When search_type = "search", an SQL expression to find the vessel of interest

ids

When search_type = "id", a vector with the vesselId of interest

includes

Enhances the response with new information, defaults to include all.

"OWNERSHIP"

returns ownership information

"AUTHORIZATIONS"

lists public authorizations for that vessel

"MATCH_CRITERIA"

adds information about the reason why a vessel is returned

match_fields

Optional. Allows to filter by matchFields levels. Possible values: "SEVERAL_FIELDS", "NO_MATCH", "ALL". Incompatible with where

registries_info_data

when search_type == "id", gets all the registry objects, only the delta or the latest.

"NONE"

The API will return the most recent object only

"DELTA"

The API will return only the objects when the vessel changed one or more identity properties

"ALL"

The registryInfo array will return the same number of objects that rows we have in the vessel database

search_type

Type of vessel search to perform. Can be "search" or "id". (Note:"advanced" and "basic" are no longer in use as of gfwr 2.0.0.)

key

Authorization token. Can be obtained with gfw_auth() function

print_request

Boolean. Whether to print the request, for debugging purposes. When contacting the GFW team it will be useful to send this string

...

Other parameters, such as limit and offset

Details

When search_type = "search" the search takes basic identity features like MMSI, IMO, callsign, shipname as inputs, using parameter "query". For more advanced SQL searches, use parameter "where". You can combine logic operators like AND, OR, =, >= , <, LIKE (for fuzzy matching). The id search allows the user to search using a GFW vessel id.

Examples

library(gfwr)
# Simple searches, using includes
get_vessel_info(query = 224224000, search_type = "search",
key = gfw_auth())
#> $dataset
#> # A tibble: 1 × 1
#>   dataset                                
#>   <chr>                                  
#> 1 public-global-vessel-identity:v20231026
#> 
#> $registryInfoTotalRecords
#> # A tibble: 1 × 1
#>   registryInfoTotalRecords
#>                      <int>
#> 1                        1
#> 
#> $registryInfo
#> # A tibble: 1 × 15
#>   id                    sourceCode ssvid flag  shipname nShipname callsign imo  
#>   <chr>                 <list>     <chr> <chr> <chr>    <chr>     <chr>    <chr>
#> 1 e0c9823749264a129d6b… <chr [6]>  2242… ESP   AGURTZA… AGURTZAB… EBSJ     8733…
#> # ℹ 7 more variables: latestVesselInfo <lgl>, transmissionDateFrom <chr>,
#> #   transmissionDateTo <chr>, geartypes <list>, lengthM <dbl>, tonnageGt <int>,
#> #   vesselInfoReference <chr>
#> 
#> $registryOwners
#> # A tibble: 2 × 6
#>   name             flag  ssvid     sourceCode dateFrom             dateTo       
#>   <chr>            <chr> <chr>     <list>     <chr>                <chr>        
#> 1 JEALSA RIANXEIRA ESP   306118000 <chr [1]>  2019-10-15T12:47:53Z 2023-09-15T1…
#> 2 JEALSA RIANXEIRA ESP   224224000 <chr [1]>  2015-10-13T16:06:33Z 2019-10-15T0…
#> 
#> $registryPublicAuthorizations
#> # A tibble: 4 × 4
#>   dateFrom             dateTo               ssvid     sourceCode
#>   <chr>                <chr>                <chr>     <list>    
#> 1 2019-10-15T00:00:00Z 2023-02-01T00:00:00Z 306118000 <chr [1]> 
#> 2 2018-01-09T00:00:00Z 2019-10-24T00:00:00Z 224224000 <chr [1]> 
#> 3 2012-01-01T00:00:00Z 2019-01-01T00:00:00Z 224224000 <chr [1]> 
#> 4 2014-03-11T00:00:00Z 2016-07-28T00:00:00Z 224224000 <chr [1]> 
#> 
#> $combinedSourcesInfo
#> # A tibble: 2 × 9
#>   vesselId  geartypes_geartype_n…¹ geartypes_geartype_s…² geartypes_geartype_y…³
#>   <chr>     <chr>                  <chr>                                   <int>
#> 1 6632c9eb… PURSE_SEINE_SUPPORT    GFW_VESSEL_LIST                          2019
#> 2 3c99c326… PURSE_SEINE_SUPPORT    GFW_VESSEL_LIST                          2015
#> # ℹ abbreviated names: ¹​geartypes_geartype_name, ²​geartypes_geartype_source,
#> #   ³​geartypes_geartype_yearFrom
#> # ℹ 5 more variables: geartypes_geartype_yearTo <int>,
#> #   shiptypes_shiptype_name <chr>, shiptypes_shiptype_source <chr>,
#> #   shiptypes_shiptype_yearFrom <int>, shiptypes_shiptype_yearTo <int>
#> 
#> $selfReportedInfo
#> # A tibble: 2 × 13
#>   vesselId         ssvid shipname nShipname flag  callsign imo   messagesCounter
#>   <chr>            <chr> <chr>    <chr>     <chr> <chr>    <chr>           <int>
#> 1 6632c9eb8-8009-… 3061… AGURTZA… AGURTZAB… BES   PJBL     8733…        21772378
#> 2 3c99c326d-dd2e-… 2242… AGURTZA… AGURTZAB… ESP   EBSJ     8733…         1887249
#> # ℹ 5 more variables: positionsCounter <int>, sourceCode <list>,
#> #   matchFields <chr>, transmissionDateFrom <chr>, transmissionDateTo <chr>
#> 
# Advanced search with where instead of query:
get_vessel_info(where = "ssvid = '441618000' OR imo = '9047271'",
search_type = "search", key = gfw_auth())
#> $dataset
#> # A tibble: 4 × 1
#>   dataset                                
#>   <chr>                                  
#> 1 public-global-vessel-identity:v20231026
#> 2 public-global-vessel-identity:v20231026
#> 3 public-global-vessel-identity:v20231026
#> 4 public-global-vessel-identity:v20231026
#> 
#> $registryInfoTotalRecords
#> # A tibble: 4 × 1
#>   registryInfoTotalRecords
#>                      <int>
#> 1                        1
#> 2                        0
#> 3                        0
#> 4                        0
#> 
#> $registryInfo
#> # A tibble: 1 × 15
#>   id                    sourceCode ssvid flag  shipname nShipname callsign imo  
#>   <chr>                 <list>     <chr> <chr> <chr>    <chr>     <chr>    <chr>
#> 1 afcdae6eacb629acda0d… <chr [9]>  4416… KOR   SEIN PH… SEINPHOE… D8SP     9047…
#> # ℹ 7 more variables: latestVesselInfo <lgl>, transmissionDateFrom <chr>,
#> #   transmissionDateTo <chr>, geartypes <list>, lengthM <dbl>, tonnageGt <int>,
#> #   vesselInfoReference <chr>
#> 
#> $registryOwners
#> # A tibble: 7 × 6
#>   name                 flag  ssvid     sourceCode dateFrom             dateTo   
#>   <chr>                <chr> <chr>     <list>     <chr>                <chr>    
#> 1 SEIN SHIPPING        KOR   441618000 <chr [1]>  2019-06-03T02:45:59Z 2024-04-…
#> 2 BAY PHOENIX SHIPPING NLD   356234000 <chr [1]>  2019-02-20T08:43:43Z 2019-06-…
#> 3 GREY SHIPPING        CUW   636018500 <chr [1]>  2018-11-12T09:19:22Z 2019-02-…
#> 4 GREY SHIPPING        CUW   3530564   <chr [1]>  2018-06-25T02:30:55Z 2018-11-…
#> 5 GREY SHIPPING        CUW   636018500 <chr [1]>  2018-02-18T06:26:20Z 2018-06-…
#> 6 GREY SHIPPING        CUW   636092423 <chr [1]>  2014-11-06T13:05:34Z 2018-02-…
#> 7 GREY SHIPPING        CUW   636015543 <chr [1]>  2014-01-01T00:04:16Z 2014-11-…
#> 
#> $registryPublicAuthorizations
#> # A tibble: 26 × 4
#>    dateFrom             dateTo               ssvid     sourceCode
#>    <chr>                <chr>                <chr>     <list>    
#>  1 2019-07-25T00:00:00Z 2024-05-01T00:00:00Z 441618000 <chr [1]> 
#>  2 2021-11-15T00:00:00Z 2024-05-01T00:00:00Z 441618000 <chr [1]> 
#>  3 2022-01-11T00:00:00Z 2024-05-01T00:00:00Z 441618000 <chr [1]> 
#>  4 2019-07-09T00:00:00Z 2024-05-01T00:00:00Z 441618000 <chr [1]> 
#>  5 2022-03-16T00:00:00Z 2023-12-31T00:00:00Z 441618000 <chr [1]> 
#>  6 2019-06-21T00:00:00Z 2020-06-01T00:00:00Z 441618000 <chr [1]> 
#>  7 2016-02-10T01:22:39Z 2019-03-19T13:38:48Z 3530564   <chr [1]> 
#>  8 2016-02-10T01:22:39Z 2019-03-19T13:38:48Z 636015543 <chr [1]> 
#>  9 2016-02-10T01:22:39Z 2019-03-19T13:38:48Z 636018500 <chr [1]> 
#> 10 2016-02-10T01:22:39Z 2019-03-19T13:38:48Z 636092423 <chr [1]> 
#> # ℹ 16 more rows
#> 
#> $combinedSourcesInfo
#> # A tibble: 9 × 9
#>   vesselId  geartypes_geartype_n…¹ geartypes_geartype_s…² geartypes_geartype_y…³
#>   <chr>     <chr>                  <chr>                                   <int>
#> 1 e11a29f7… CARRIER                GFW_VESSEL_LIST                          2012
#> 2 20345d8c… CARRIER                GFW_VESSEL_LIST                          2018
#> 3 1f10e42b… CARRIER                GFW_VESSEL_LIST                          2019
#> 4 14e3a7e8… CARRIER                GFW_VESSEL_LIST                          2019
#> 5 30c6ba6a… CARRIER                GFW_VESSEL_LIST                          2014
#> 6 b5800873… CARRIER                GFW_VESSEL_LIST                          2018
#> 7 77fabfb1… OTHER                  COMBINATION_OF_REGIST…                   2013
#> 8 b2e595b6… CARGO                  COMBINATION_OF_REGIST…                   2019
#> 9 559fa4cf… OTHER                  COMBINATION_OF_REGIST…                   2012
#> # ℹ abbreviated names: ¹​geartypes_geartype_name, ²​geartypes_geartype_source,
#> #   ³​geartypes_geartype_yearFrom
#> # ℹ 5 more variables: geartypes_geartype_yearTo <int>,
#> #   shiptypes_shiptype_name <chr>, shiptypes_shiptype_source <chr>,
#> #   shiptypes_shiptype_yearFrom <int>, shiptypes_shiptype_yearTo <int>
#> 
#> $selfReportedInfo
#> # A tibble: 9 × 13
#>   vesselId         ssvid shipname nShipname flag  callsign imo   messagesCounter
#>   <chr>            <chr> <chr>    <chr>     <chr> <chr>    <chr>           <int>
#> 1 1f10e42bc-cfbb-… 4416… SEIN PH… SEINPHOE… KOR   D8SP     9047…        90844256
#> 2 14e3a7e81-1e4e-… 4416… SEIN PH… SEINPHOE… KOR   D8SP     9047…         3199467
#> 3 20345d8c6-692a-… 6360… BAY  PH… BAYPHOEN… LBR   A8OE5    9047…          857880
#> 4 b58008738-854b-… 3530… BAY  PH… BAYPHOEN… PAN   A8OE5    9047…          446403
#> 5 30c6ba6ac-c6c3-… 6360… BAY  PH… BAYPHOEN… LBR   A8OE5    9047…         9984723
#> 6 e11a29f7b-b056-… 6360… BAY  PH… BAYPHOEN… LBR   A8OE5    9047…         1515842
#> 7 77fabfb1d-d3ea-… 8442… BAY  PH… BAYPHOEN… NA    A8OE5    9047…          150814
#> 8 b2e595b63-3a8e-… 3562… BAY PHO… BAYPHOEN… PAN   3FCZ     9047…         2638091
#> 9 559fa4cf4-467e-… 6360… SUMMER … SUMMERPH… LBR   A8OE5    9047…           19655
#> # ℹ 5 more variables: positionsCounter <int>, sourceCode <list>,
#> #   matchFields <chr>, transmissionDateFrom <chr>, transmissionDateTo <chr>
#> 
 # Vessel id search
 get_vessel_info(search_type = "id",
 ids = c("8c7304226-6c71-edbe-0b63-c246734b3c01",
 "6583c51e3-3626-5638-866a-f47c3bc7ef7c"), key = gfw_auth())
#> $dataset
#> # A tibble: 2 × 1
#>   dataset                                
#>   <chr>                                  
#> 1 public-global-vessel-identity:v20231026
#> 2 public-global-vessel-identity:v20231026
#> 
#> $registryInfoTotalRecords
#> # A tibble: 2 × 1
#>   registryInfoTotalRecords
#>                      <int>
#> 1                        2
#> 2                        1
#> 
#> $registryInfo
#> # A tibble: 3 × 15
#>   id                    sourceCode ssvid flag  shipname nShipname callsign imo  
#>   <chr>                 <list>     <chr> <chr> <chr>    <chr>     <chr>    <chr>
#> 1 a8d00ce54b37add7f85a… <chr [6]>  2106… CYP   FRIO FO… FRIOFORW… 5BWC3    9076…
#> 2 a8d00ce54b37add7f85a… <chr [2]>  2733… RUS   FRIO FO… FRIOFORW… UCRZ     9076…
#> 3 b82d02e5c2c11e5fe536… <chr [5]>  4417… KOR   ADRIA    ADRIA     DTBY3    8919…
#> # ℹ 7 more variables: latestVesselInfo <lgl>, transmissionDateFrom <chr>,
#> #   transmissionDateTo <chr>, geartypes <list>, lengthM <dbl>, tonnageGt <dbl>,
#> #   vesselInfoReference <chr>
#> 
#> $registryOwners
#> # A tibble: 3 × 6
#>   name               flag  ssvid     sourceCode dateFrom             dateTo     
#>   <chr>              <chr> <chr>     <list>     <chr>                <chr>      
#> 1 COLINER            CYP   210631000 <chr [1]>  2014-01-01T00:16:58Z 2024-04-30…
#> 2 COLINER            CYP   273379740 <chr [1]>  2015-02-27T10:59:43Z 2018-03-21…
#> 3 DONGWON INDUSTRIES KOR   441734000 <chr [2]>  2014-01-18T19:46:21Z 2024-03-22…
#> 
#> $registryPublicAuthorizations
#> # A tibble: 4 × 4
#>   dateFrom             dateTo               ssvid     sourceCode
#>   <chr>                <chr>                <chr>     <list>    
#> 1 2022-12-19T00:00:00Z 2024-05-01T00:00:00Z 210631000 <chr [1]> 
#> 2 2020-01-01T00:00:00Z 2024-05-01T00:00:00Z 210631000 <chr [1]> 
#> 3 2013-09-20T00:00:00Z 2024-05-01T00:00:00Z 441734000 <chr [1]> 
#> 4 2015-10-08T00:00:00Z 2020-07-21T00:00:00Z 441734000 <chr [1]> 
#> 
#> $combinedSourcesInfo
#> # A tibble: 6 × 9
#>   vesselId  geartypes_geartype_n…¹ geartypes_geartype_s…² geartypes_geartype_y…³
#>   <chr>     <chr>                  <chr>                                   <int>
#> 1 da1cd7e1… CARRIER                GFW_VESSEL_LIST                          2015
#> 2 8c730422… CARRIER                GFW_VESSEL_LIST                          2013
#> 3 6583c51e… OTHER                  COMBINATION_OF_REGIST…                   2013
#> 4 6583c51e… OTHER                  COMBINATION_OF_REGIST…                   2013
#> 5 6583c51e… TUNA_PURSE_SEINES      COMBINATION_OF_REGIST…                   2014
#> 6 6583c51e… TUNA_PURSE_SEINES      COMBINATION_OF_REGIST…                   2014
#> # ℹ abbreviated names: ¹​geartypes_geartype_name, ²​geartypes_geartype_source,
#> #   ³​geartypes_geartype_yearFrom
#> # ℹ 5 more variables: geartypes_geartype_yearTo <int>,
#> #   shiptypes_shiptype_name <chr>, shiptypes_shiptype_source <chr>,
#> #   shiptypes_shiptype_yearFrom <int>, shiptypes_shiptype_yearTo <int>
#> 
#> $selfReportedInfo
#> # A tibble: 2 × 13
#>   vesselId         ssvid shipname nShipname flag  callsign imo   messagesCounter
#>   <chr>            <chr> <chr>    <chr>     <chr> <chr>    <chr>           <int>
#> 1 8c7304226-6c71-… 2106… FRIO FO… FRIOFORW… CYP   5BWC3    9076…       263134677
#> 2 6583c51e3-3626-… 4417… ADRIA    ADRIA     KOR   DTBY3    8919…         3742574
#> # ℹ 5 more variables: positionsCounter <int>, sourceCode <list>,
#> #   matchFields <chr>, transmissionDateFrom <chr>, transmissionDateTo <chr>
#> 
 all <- get_vessel_info(search_type = "id",
 ids = c("8c7304226-6c71-edbe-0b63-c246734b3c01"),
 registries_info_data = c("ALL"), key = gfw_auth())
 none <- get_vessel_info(search_type = "id",
 ids = c("8c7304226-6c71-edbe-0b63-c246734b3c01"),
 registries_info_data = c("NONE"), key = gfw_auth())
 delta <- get_vessel_info(search_type = "id",
 ids = c("8c7304226-6c71-edbe-0b63-c246734b3c01"),
 registries_info_data = c("DELTA"), key = gfw_auth())