gfwr: Access data from Global Fishing Watch APIs

This package connects to several Global Fishing Watch APIs to get vessel and events information in an R-friendly format.

See also

Author

Maintainer: Andrea Sánchez-Tapia andrea.sanchez-tapia@globalfishingwatch.org (ORCID)

Authors:

Tyler Clavelle tyler@globalfishingwatch.org (ORCID)

Rocío Joo rocio.joo@globalfishingwatch.org (ORCID)

Nate Miller nate@globalfishingwatch.org

Jorge Cornejo-Donoso jorge.cornejo@globalfishingwatch.org (ORCID)