Base function to get events from API and convert response to data frame

Base function to get events from API and convert response to data frame

get_event(
  event_type,
  encounter_types = NULL,
  vessels = NULL,
  flags = NULL,
  vessel_types = NULL,
  start_date = "2012-01-01",
  end_date = "2024-12-31",
  region = NULL,
  region_source = NULL,
  gap_intentional_disabling = NULL,
  duration = 1,
  confidences = c(2, 3, 4),
  limit = 99999,
  offset = 0,
  sort = "+start",
  key = gfw_auth(),
  quiet = FALSE,
  print_request = FALSE,
  ...
)

Arguments

event_type

Type of event to get data of. A vector with any combination of "ENCOUNTER", "FISHING", "GAP", "LOITERING", "PORT_VISIT"

encounter_types

Filters for types of vessels during the encounter. A vector with any combination of: "CARRIER-FISHING", "FISHING-CARRIER", "FISHING-SUPPORT", "SUPPORT-FISHING"

vessels

A vector of VesselIDs, obtained via the get_vessel_info() function

flags

ISO3 code for the flag of the vessels. Null by default.

vessel_types

A vector of vessel types, any combination of: "FISHING", "CARRIER", "SUPPORT", "PASSENGER", "OTHER_NON_FISHING", "SEISMIC_VESSEL", "BUNKER_OR_TANKER", "CARGO"

start_date

Start of date range to search events, in YYYY-MM-DD format and including this date

end_date

End of date range to search events, in YYYY-MM-DD format and excluding this date

region

geojson shape to filter raster or GFW region code (such as an EEZ code). See details about formatting the geojson

region_source

source of the region ('EEZ','MPA', 'RFMO' or 'USER_JSON')

gap_intentional_disabling

Logical. Whether the Gap events are intentional, according to Global Fishing Watch algorithms

duration

duration, in minutes, of the event, ex. 30

confidences

Confidence levels (1-4) of events (port visits only)

limit

Limit

offset

Offset

sort

How to sort the events. By default, +start, which sorts the events in ascending order (+) of the start dates of the events. Other possible values are -start, +end, -end.

key

Authorization token. Can be obtained with gfw_auth() function

quiet

Boolean. Whether to print the number of events returned by the request

print_request

Boolean. Whether to print the request, for debugging purposes. When contacting the GFW team it will be useful to send this string

...

Other arguments

Details

There are currently four available event types and these events are provided for three vessel types - fishing, carrier, and support vessels. Fishing events (event_type = "FISHING") are specific to fishing vessels and loitering events (event_type = "LOITERING") are specific to carrier vessels. Port visits (event_type = "PORT_VISIT") and encounters (event_type = "ENCOUNTER") are available for all vessel types. For more details about the various event types, see the GFW API documentation.

Encounter events involve multiple vessels and one row is returned for each vessel involved in an encounter. For example, an encounter between a carrier and fishing vessel (CARRIER-FISHING) will have one row for the fishing vessel and one for the carrier vessel. The id field for encounter events has two components separated by a .. The first component is the unique id for the encounter event and will be the same for all vessels involved in the encounter. The second component is an integer used to distinguish between different vessels in the encounter.

Examples

library(gfwr)
# port visits
get_event(event_type = "PORT_VISIT",
          vessels = c("e0c9823749264a129d6b47a7aabce377",
          "8c7304226-6c71-edbe-0b63-c246734b3c01"),
          start_date = "2017-01-26",
          end_date = "2017-12-31",
          confidence = c(3, 4), # only for port visits
          key = gfw_auth())
#> [1] "Downloading 38 events from GFW"
#> # A tibble: 38 × 14
#>    start               end                 id    type    lat    lon regions     
#>    <dttm>              <dttm>              <chr> <chr> <dbl>  <dbl> <list>      
#>  1 2016-12-18 19:42:49 2017-03-18 06:44:41 9d58… port…  52.5   4.60 <named list>
#>  2 2017-03-20 12:57:50 2017-03-31 11:54:21 7c83… port…  62.1  -6.71 <named list>
#>  3 2017-04-05 09:13:14 2017-04-06 00:39:44 d7ed… port…  60.0  29.4  <named list>
#>  4 2017-04-06 02:30:14 2017-04-09 21:59:37 3bf9… port…  59.9  30.2  <named list>
#>  5 2017-04-16 11:23:45 2017-04-25 06:07:19 36d3… port…  62.1  -6.72 <named list>
#>  6 2017-04-30 02:20:02 2017-05-04 19:13:25 b2d7… port…  59.9  30.2  <named list>
#>  7 2017-05-10 20:31:07 2017-06-03 07:16:36 466e… port…  64.1 -22.1  <named list>
#>  8 2017-06-07 18:27:05 2017-06-08 18:04:45 5d7a… port…  52.5   4.12 <named list>
#>  9 2017-06-08 19:44:51 2017-06-14 18:54:26 1212… port…  52.5   4.60 <named list>
#> 10 2017-06-18 16:50:57 2017-06-19 02:04:34 d070… port…  60.0  29.4  <named list>
#> # ℹ 28 more rows
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
 #encounters
 get_event(event_type = "ENCOUNTER",
 vessels = c("e0c9823749264a129d6b47a7aabce377",
  "8c7304226-6c71-edbe-0b63-c246734b3c01"),
  start_date = "2018-01-30",
  end_date = "2023-02-04",
  key = gfw_auth())
#> [1] "Your request returned zero results"
#> NULL
 # fishing
 get_event(event_type = "FISHING",
 vessels = c("8c7304226-6c71-edbe-0b63-c246734b3c01"),
  start_date = "2017-01-26",
  end_date = "2023-02-04",
  key = gfw_auth())
#> [1] "Your request returned zero results"
#> NULL
 # GAPS
 get_event(event_type = "GAP",
 vessels = c("e0c9823749264a129d6b47a7aabce377",
  "8c7304226-6c71-edbe-0b63-c246734b3c01"),
  start_date = "2017-01-26",
  end_date = "2023-02-04",
  key = gfw_auth())
#> [1] "Downloading 1 events from GFW"
#> # A tibble: 1 × 14
#>   start               end                 id      type    lat   lon regions     
#>   <dttm>              <dttm>              <chr>   <chr> <dbl> <dbl> <list>      
#> 1 2019-07-30 19:41:55 2019-07-31 07:58:34 1dd8d9… gap    2.72 -6.32 <named list>
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
 # loitering
 get_event(event_type = "LOITERING",
 vessels = c("e0c9823749264a129d6b47a7aabce377",
  "8c7304226-6c71-edbe-0b63-c246734b3c01"),
  start_date = "2017-01-26",
  end_date = "2023-02-04",
  key = gfw_auth())
#> [1] "Downloading 48 events from GFW"
#> # A tibble: 48 × 14
#>    start               end                 id    type    lat    lon regions     
#>    <dttm>              <dttm>              <chr> <chr> <dbl>  <dbl> <list>      
#>  1 2017-04-03 20:58:35 2017-04-04 03:52:35 4519… loit… 57.1   20.0  <named list>
#>  2 2018-11-20 16:48:51 2018-11-21 11:49:35 e74c… loit… 23.7  -16.5  <named list>
#>  3 2018-11-21 16:25:20 2018-11-23 10:58:04 cf93… loit… 23.6  -16.4  <named list>
#>  4 2018-11-23 15:40:50 2018-11-24 12:59:05 8b51… loit… 23.7  -16.5  <named list>
#>  5 2019-01-01 19:58:14 2019-01-01 20:49:42 a466… loit…  3.63 -12.7  <named list>
#>  6 2019-01-01 21:44:55 2019-01-01 23:54:39 8353… loit…  3.59 -12.7  <named list>
#>  7 2019-02-22 17:57:11 2019-02-22 21:59:52 3366… loit…  2.97   2.70 <named list>
#>  8 2019-02-22 23:59:03 2019-02-23 01:58:58 484e… loit…  2.97   2.76 <named list>
#>  9 2019-03-23 11:45:06 2019-03-23 14:54:02 3bcc… loit…  3.49  -2.64 <named list>
#> 10 2019-04-27 15:45:49 2019-04-28 00:57:07 c11d… loit…  1.31  -1.49 <named list>
#> # ℹ 38 more rows
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
 # encounter type
 get_event(event_type = "ENCOUNTER",
 encounter_types = "CARRIER-FISHING",
 start_date = "2020-01-01",
 end_date = "2020-01-31",
 key = gfw_auth())
#> [1] "Downloading 791 events from GFW"
#> # A tibble: 791 × 14
#>    start               end                 id                type     lat    lon
#>    <dttm>              <dttm>              <chr>             <chr>  <dbl>  <dbl>
#>  1 2019-12-31 08:40:00 2020-01-01 07:40:00 a3cff76a070a919f… enco…  57.5   157. 
#>  2 2019-12-31 12:00:00 2020-01-01 13:50:00 b059d20534c7fd5f… enco… -17.6   -79.3
#>  3 2019-12-31 12:50:00 2020-01-01 09:50:00 cd07d7e5d65e81b3… enco… -17.7   -79.2
#>  4 2019-12-31 14:50:00 2020-01-01 20:20:00 13dac0526c993292… enco… -17.6   -79.4
#>  5 2019-12-31 16:00:00 2020-01-01 08:50:00 2e8b8040d87ad0ae… enco…  -3.44 -147. 
#>  6 2019-12-31 17:50:00 2020-01-01 16:10:00 d374df02769d8d55… enco…  68.7    50.3
#>  7 2019-12-31 17:50:00 2020-01-01 03:00:00 c1f7885ca82db1b3… enco… -17.9   -79.1
#>  8 2020-01-01 00:00:00 2020-01-02 15:40:00 7286e857b8f926bb… enco… -17.5   -79.4
#>  9 2020-01-01 00:00:00 2020-01-01 15:10:00 bb6754d7fb452156… enco…  60.2   166. 
#> 10 2020-01-01 00:00:00 2020-01-01 04:30:00 463f0c8e2200beae… enco… -45.6   -60.4
#> # ℹ 781 more rows
#> # ℹ 8 more variables: regions <list>, eez <list>, rfmo <list>, fao <list>,
#> #   boundingBox <list>, distances <list>, vessel <list>, event_info <list>
 # vessel types
 get_event(event_type = "ENCOUNTER",
 vessel_types = c("CARRIER", "FISHING"),
 start_date = "2020-01-01",
 end_date = "2020-01-31",
 key = gfw_auth())
#> [1] "Downloading 5112 events from GFW"
#> # A tibble: 5,112 × 14
#>    start               end                 id    type    lat    lon regions     
#>    <dttm>              <dttm>              <chr> <chr> <dbl>  <dbl> <list>      
#>  1 2019-12-17 14:10:00 2020-01-02 04:10:00 8c07… enco… 67.5    15.5 <named list>
#>  2 2019-12-17 14:10:00 2020-01-02 04:10:00 8c07… enco… 67.5    15.5 <named list>
#>  3 2019-12-26 00:20:00 2020-01-07 23:50:00 59d4… enco… -1.82 -113.  <named list>
#>  4 2019-12-26 00:20:00 2020-01-07 23:50:00 59d4… enco… -1.82 -113.  <named list>
#>  5 2019-12-26 14:10:00 2020-01-03 05:30:00 60c1… enco… -1.79 -113.  <named list>
#>  6 2019-12-26 14:10:00 2020-01-03 05:30:00 60c1… enco… -1.79 -113.  <named list>
#>  7 2019-12-27 09:10:00 2020-01-06 14:00:00 2159… enco…  9.50  -99.1 <named list>
#>  8 2019-12-27 09:10:00 2020-01-06 14:00:00 2159… enco…  9.50  -99.1 <named list>
#>  9 2019-12-30 02:20:00 2020-01-13 05:40:00 87de… enco… -1.84 -111.  <named list>
#> 10 2019-12-30 02:20:00 2020-01-13 05:40:00 87de… enco… -1.84 -111.  <named list>
#> # ℹ 5,102 more rows
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
# fishing events in user geojson
region <- '"geometry": {"type": "Polygon","coordinates":
[[[120.36621093749999,26.725986812271756],[122.36572265625,26.725986812271756],
[122.36572265625,28.323724553546015],[120.36621093749999,28.323724553546015],
[120.36621093749999,26.725986812271756]]]}'
get_event(event_type = 'FISHING',
              start_date = "2017-01-01",
              end_date = "2017-01-31",
              region = region,
              region_source = 'USER_JSON',
              flags = 'CHN',
              key = gfw_auth())
#> [1] "Downloading 2480 events from GFW"
#> # A tibble: 2,480 × 14
#>    start               end                 id     type    lat   lon regions     
#>    <dttm>              <dttm>              <chr>  <chr> <dbl> <dbl> <list>      
#>  1 2016-12-30 07:28:06 2017-01-02 03:49:40 8326e… fish…  27.1  121. <named list>
#>  2 2016-12-30 07:46:42 2017-01-02 05:31:09 b599a… fish…  27.1  121. <named list>
#>  3 2016-12-30 07:46:46 2017-01-01 08:48:16 ee787… fish…  27.3  121. <named list>
#>  4 2016-12-30 07:47:06 2017-01-01 12:52:33 eceae… fish…  27.1  121. <named list>
#>  5 2016-12-30 07:48:18 2017-01-02 02:11:48 28f64… fish…  27.3  121. <named list>
#>  6 2016-12-30 07:48:54 2017-01-01 09:27:53 7b9e1… fish…  27.4  121. <named list>
#>  7 2016-12-30 07:49:02 2017-01-02 03:05:48 07083… fish…  27.2  121. <named list>
#>  8 2016-12-30 09:24:30 2017-01-02 03:06:00 4ae5c… fish…  27.3  121. <named list>
#>  9 2016-12-30 09:24:38 2017-01-01 00:15:49 bc1dd… fish…  27.1  121. <named list>
#> 10 2016-12-30 09:25:46 2017-01-01 13:02:49 10e51… fish…  27.2  121. <named list>
#> # ℹ 2,470 more rows
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
# fishing events in Senegal EEZ
get_event(event_type = 'FISHING',
              start_date = "2020-10-01",
              end_date = "2020-12-31",
              region = 8371,
              region_source = 'EEZ',
              flags = 'CHN',
              key = gfw_auth())
#> [1] "Downloading 832 events from GFW"
#> # A tibble: 832 × 14
#>    start               end                 id     type    lat   lon regions     
#>    <dttm>              <dttm>              <chr>  <chr> <dbl> <dbl> <list>      
#>  1 2020-09-30 08:26:23 2020-10-01 00:35:22 4960d… fish…  14.7 -17.4 <named list>
#>  2 2020-09-30 17:16:14 2020-10-01 04:38:44 d1d6a… fish…  14.7 -17.4 <named list>
#>  3 2020-10-01 11:11:21 2020-10-01 16:58:03 c3918… fish…  12.8 -17.4 <named list>
#>  4 2020-10-01 12:20:20 2020-10-01 18:56:21 a8ef8… fish…  14.7 -17.4 <named list>
#>  5 2020-10-01 13:33:31 2020-10-02 05:49:12 689f2… fish…  14.7 -17.4 <named list>
#>  6 2020-10-01 21:32:55 2020-10-02 08:21:07 ba95f… fish…  12.9 -17.1 <named list>
#>  7 2020-10-02 13:37:11 2020-10-02 16:08:52 bce17… fish…  12.3 -17.0 <named list>
#>  8 2020-10-02 14:29:01 2020-10-02 17:05:05 d70e4… fish…  12.3 -17.0 <named list>
#>  9 2020-10-02 15:34:28 2020-10-02 22:28:01 91ad9… fish…  13.6 -17.1 <named list>
#> 10 2020-10-02 22:15:13 2020-10-03 01:47:01 27d55… fish…  12.3 -17.0 <named list>
#> # ℹ 822 more rows
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
# port visits in user region by ATG vessels
region = '"geometry":{"type": "Polygon","coordinates":
[[[30.552978515625,46.255846818480315],[31.22314453125,46.255846818480315],
[31.22314453125,46.59661864884465],[30.552978515625,46.59661864884465],
[30.552978515625,46.255846818480315]]]}'
get_event(event_type = 'PORT_VISIT',
         vessels = "e0248aed9-99b4-bae7-6b87-ff0a3c464676",
         start_date = "2017-01-01",
         end_date = "2017-01-31",
         region = region,
         region_source = 'USER_JSON',
         flags = 'ATG',
         duration = 60)
#> [1] "Downloading 1 events from GFW"
#> # A tibble: 1 × 14
#>   start               end                 id      type    lat   lon regions     
#>   <dttm>              <dttm>              <chr>   <chr> <dbl> <dbl> <list>      
#> 1 2017-01-01 00:00:00 2017-01-01 09:35:02 4f76e3… port…  46.3  31.0 <named list>
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
# loitering events in user region by KOR vessels
region = '"geometry": {"type": "Polygon","coordinates":
[[[43.835981972515576,-6.785011952437713],[43.83602857589722,-6.785011952437713],
[43.83602857589722,-6.784984652340707],[43.835981972515576,-6.784984652340707],
[43.835981972515576,-6.785011952437713]]]}'
get_event(event_type = 'LOITERING',
                          vessels = "82be6f228-8ce4-26d1-bf81-3b7979d0c72f",
                          start_date = "2017-01-01",
                          end_date = "2017-01-31",
                          region = region,
                          region_source = 'USER_JSON',
                          flags = 'KOR',
                          duration = 60)
#> [1] "Your request returned zero results"
#> NULL
# encounter events in user region by TWN vessels
region = '"geometry": {"type": "Polygon","coordinates":
[[[-130.9735107421875,-17.691128657307427],[-130.4901123046875,-17.691128657307427],
[-130.4901123046875,-17.209017141391765],[-130.9735107421875,-17.209017141391765],
[-130.9735107421875,-17.691128657307427]]]}'
get_event(event_type = 'ENCOUNTER',
         vessels = "55d38c0ee-e0d7-cb32-ac9c-8b3680d213b3",
         start_date = "2017-01-01",
         end_date = "2017-01-31",
         region = region,
         region_source = 'USER_JSON',
         flags = 'TWN',
         duration = 60)
#> [1] "Downloading 1 events from GFW"
#> # A tibble: 1 × 14
#>   start               end                 id      type    lat   lon regions     
#>   <dttm>              <dttm>              <chr>   <chr> <dbl> <dbl> <list>      
#> 1 2017-01-01 00:50:00 2017-01-01 04:30:00 05e5fc… enco… -17.5 -131. <named list>
#> # ℹ 7 more variables: eez <list>, rfmo <list>, fao <list>, boundingBox <list>,
#> #   distances <list>, vessel <list>, event_info <list>
get_event(event_type = 'GAP',
         start_date = "2017-01-01",
         end_date = "2017-01-31",
         flags = 'TWN',
         gap_intentional_disabling = TRUE)
#> [1] "Downloading 195 events from GFW"
#> # A tibble: 195 × 14
#>    start               end                 id               type      lat    lon
#>    <dttm>              <dttm>              <chr>            <chr>   <dbl>  <dbl>
#>  1 2017-01-01 02:34:58 2017-01-01 16:31:39 ebbcf21c1e4fb5b… gap   -14.1   -174. 
#>  2 2017-01-01 05:28:06 2017-01-04 06:44:45 57dd3aa4d80a807… gap    -3       50.4
#>  3 2017-01-01 05:38:22 2017-01-01 18:28:45 5e5e7af43895d05… gap   -22.5     50.5
#>  4 2017-01-01 08:30:13 2017-01-08 18:31:11 5e4951178c6c8bd… gap    -0.616 -180. 
#>  5 2017-01-01 12:28:58 2017-01-04 07:05:02 e3ad85fd0710127… gap    -2.77  -165. 
#>  6 2017-01-01 21:31:45 2017-01-02 11:15:04 7c819706694c925… gap    -2.92  -165. 
#>  7 2017-01-01 22:23:23 2017-01-02 16:29:48 c1c2e23c70e42ad… gap    -7.57    54.9
#>  8 2017-01-02 01:56:28 2017-01-05 23:11:55 dce8f7f0fae1c2f… gap   -21.6     61.3
#>  9 2017-01-02 02:07:46 2017-01-02 15:36:19 f523cdb0c4d3d3d… gap   -14.1   -174. 
#> 10 2017-01-02 03:49:47 2017-01-03 18:58:02 79cdd1599e2c08c… gap    -2.13   163. 
#> # ℹ 185 more rows
#> # ℹ 8 more variables: regions <list>, eez <list>, rfmo <list>, fao <list>,
#> #   boundingBox <list>, distances <list>, vessel <list>, event_info <list>