Base function to get raster from API and convert response to data frame

Base function to get raster from API and convert response to data frame

get_raster(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  group_by = NULL,
  filter_by = NULL,
  start_date = NULL,
  end_date = NULL,
  region = NULL,
  region_source = NULL,
  key = gfw_auth(),
  print_request = FALSE
)

Arguments

spatial_resolution

raster spatial resolution. Can be "LOW" = 0.1 degree or "HIGH" = 0.01 degree

temporal_resolution

raster temporal resolution. Can be 'HOURLY', 'DAILY', 'MONTHLY', 'YEARLY'

group_by

parameter to group by. Can be 'VESSEL_ID', 'FLAG', 'GEARTYPE', 'FLAGANDGEARTYPE' or 'MMSI'. Optional.

filter_by

parameter to filter by.

start_date

Start of date range to search events, in YYYY-MM-DD format and including this date

end_date

End of date range to search events, in YYYY-MM-DD format and excluding this date

region

sf shape to filter raster or GFW region code (such as a Marine Regions Geographic Identifier or EEZ code).

region_source

source of the region ('EEZ','MPA', 'RFMO' or 'USER_JSON')

key

Authorization token. Can be obtained with gfw_auth() function

print_request

Boolean. Whether to print the request, for debugging purposes. When contacting the GFW team it will be useful to send this string

Examples

library(gfwr)
# using region codes
code_eez <- get_region_id(region_name = 'CIV', region_source = 'EEZ',
key = gfw_auth())
get_raster(spatial_resolution = 'LOW',
           temporal_resolution = 'YEARLY',
           group_by = 'FLAG',
           start_date = "2021-01-01",
           end_date = "2021-10-01",
           region = code_eez$id,
           region_source = 'EEZ',
           key = gfw_auth(),
           print_request = TRUE)
#> <httr2_request>
#> POST
#> https://gateway.api.globalfishingwatch.org/v3/4wings/report?datasets%5B0%5D=public-global-fishing-effort%3Alatest&spatial-resolution=LOW&temporal-resolution=YEARLY&group-by=FLAG&date-range=2021-01-01%2C2021-10-01&format=CSV
#> Headers:
#> • Authorization: 'Bearer
#> eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImtpZEtleSJ9.eyJkYXRhIjp7Im5hbWUiOiJnZndyIiwidXNlcklkIjoyNjEyMywiYXBwbGljYXRpb25OYW1lIjoiZ2Z3ciIsImlkIjo2OTgsInR5cGUiOiJ1c2VyLWFwcGxpY2F0aW9uIn0sImlhdCI6MTY4NzQ4MzY4MywiZXhwIjoyMDAyODQzNjgzLCJhdWQiOiJnZnciLCJpc3MiOiJnZncifQ.dRaU2jUYVaOKMwCcAJA_6brmBwForrNMxWWaxKXPl8WvB353rnFTZ2nM2BqotdvLXyLMaZkR5LopRd4BVx2fWX7-vJPyE1DepOAOZwq2eaRxfE8t-7Kutb9wPUuZaVOGgLRn2IjMt07zUBoe5VPJ3BddoLpjL5c796w4D97dPfsZ70RQMApxChPZAKfp_Qh3RjC_K8QBr6fMQSioWOoNb8g9uUJTjfGBVOJ0VL_pSBi_y-n0S5Z32kTRbX7XmaEbiGDEWJHqLf6-fNtoAnqpB-FPGtcMjtIU2ol06qXJ6jOTiwL2kEWjNS9ExNRxX166ch-wYNGgyUzQYj-rRuV_LJrRahdxXppcTEH8yawjkaLOWqMS0rNWH_VZJglh91ZxuiT15R1lDmg36e4OEsGJzMjRYN5YLxyHXRBBnBhNPxt7txafO7M5vDajCrjQDi-JyqgDKwizT1HbuG9Y0QMaRQ4th4F2RRE3u-JAu1nHVi8Ikr4s4WmYt1GpQRfGZuPn'
#> • Content-Type: 'application/json'
#> Body: a string
#> Rows: 611 Columns: 6
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): flag
#> dbl (5): Lat, Lon, Time Range, Vessel IDs, Apparent Fishing Hours
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 611 × 6
#>      Lat   Lon `Time Range` flag  `Vessel IDs` `Apparent Fishing Hours`
#>    <dbl> <dbl>        <dbl> <chr>        <dbl>                    <dbl>
#>  1   4.9  -5.8         2021 CHN              4                    31   
#>  2   4.9  -5.9         2021 CHN              3                    11.7 
#>  3   5.2  -4           2021 GTM              2                     2.58
#>  4   2.7  -5.3         2021 FRA              1                     0.25
#>  5   2.7  -4           2021 ESP              1                     6.06
#>  6   1.9  -4.2         2021 FRA              1                     2.21
#>  7   1.9  -4.1         2021 FRA              1                     0.04
#>  8   4    -6.9         2021 FRA              1                     2.99
#>  9   4.3  -4.1         2021 FRA              2                     3.51
#> 10   4.7  -6.2         2021 CHN              2                    12.9 
#> # ℹ 601 more rows
#using a sf from disk /loading a test sf object
data(test_shape)
get_raster(spatial_resolution = 'LOW',
            temporal_resolution = 'YEARLY',
            start_date = '2021-01-01',
            end_date = '2021-10-01',
            region = test_shape,
            region_source = 'USER_JSON',
            key = gfw_auth(),
            print_request = TRUE)
#> <httr2_request>
#> POST
#> https://gateway.api.globalfishingwatch.org/v3/4wings/report?datasets%5B0%5D=public-global-fishing-effort%3Alatest&spatial-resolution=LOW&temporal-resolution=YEARLY&date-range=2021-01-01%2C2021-10-01&format=CSV
#> Headers:
#> • Authorization: 'Bearer
#> eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImtpZEtleSJ9.eyJkYXRhIjp7Im5hbWUiOiJnZndyIiwidXNlcklkIjoyNjEyMywiYXBwbGljYXRpb25OYW1lIjoiZ2Z3ciIsImlkIjo2OTgsInR5cGUiOiJ1c2VyLWFwcGxpY2F0aW9uIn0sImlhdCI6MTY4NzQ4MzY4MywiZXhwIjoyMDAyODQzNjgzLCJhdWQiOiJnZnciLCJpc3MiOiJnZncifQ.dRaU2jUYVaOKMwCcAJA_6brmBwForrNMxWWaxKXPl8WvB353rnFTZ2nM2BqotdvLXyLMaZkR5LopRd4BVx2fWX7-vJPyE1DepOAOZwq2eaRxfE8t-7Kutb9wPUuZaVOGgLRn2IjMt07zUBoe5VPJ3BddoLpjL5c796w4D97dPfsZ70RQMApxChPZAKfp_Qh3RjC_K8QBr6fMQSioWOoNb8g9uUJTjfGBVOJ0VL_pSBi_y-n0S5Z32kTRbX7XmaEbiGDEWJHqLf6-fNtoAnqpB-FPGtcMjtIU2ol06qXJ6jOTiwL2kEWjNS9ExNRxX166ch-wYNGgyUzQYj-rRuV_LJrRahdxXppcTEH8yawjkaLOWqMS0rNWH_VZJglh91ZxuiT15R1lDmg36e4OEsGJzMjRYN5YLxyHXRBBnBhNPxt7txafO7M5vDajCrjQDi-JyqgDKwizT1HbuG9Y0QMaRQ4th4F2RRE3u-JAu1nHVi8Ikr4s4WmYt1GpQRfGZuPn'
#> • Content-Type: 'application/json'
#> Body: a string
#> Rows: 6524 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl (5): Lat, Lon, Time Range, Vessel IDs, Apparent Fishing Hours
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 6,524 × 5
#>      Lat   Lon `Time Range` `Vessel IDs` `Apparent Fishing Hours`
#>    <dbl> <dbl>        <dbl>        <dbl>                    <dbl>
#>  1  17.1  61.1         2021            6                    65.4 
#>  2  16.8  60.8         2021            1                     4.5 
#>  3  17.2  61.2         2021            5                    30.5 
#>  4  17.3  61.5         2021            1                     8.77
#>  5  17    61.2         2021            2                    18.5 
#>  6  16.8  61.5         2021            1                     5.89
#>  7  18.1  63           2021            2                    12.0 
#>  8  17.1  62.2         2021            7                    63.8 
#>  9  17    62.2         2021            5                    60.0 
#> 10  17    62.3         2021            7                    45   
#> # ℹ 6,514 more rows