Base function to get raster from API and convert response to data frame

Base function to get raster from API and convert response to data frame

get_raster(
  spatial_resolution = NULL,
  temporal_resolution = NULL,
  group_by = NULL,
  filter_by = NULL,
  date_range = NULL,
  region = NULL,
  region_source = NULL,
  key = gfw_auth()
)

Arguments

spatial_resolution

raster spatial resolution. Can be "LOW" = 0.1 degree or "HIGH" = 0.01 degree

temporal_resolution

raster temporal resolution. Can be 'HOURLY', 'DAILY', 'MONTHLY', 'YEARLY'

group_by

parameter to group by. Can be 'VESSEL_ID', 'FLAG', 'GEARTYPE', 'FLAGANDGEARTYPE' or 'MMSI'

filter_by

parameter to filter by.

date_range

Start and end of date range for raster (must be 366 days or less). Formatted "YYYY-MM-DD,YYYY-MM-DD"

region

geojson shape to filter raster or GFW region code (such as an EEZ code). See details about formatting the geojson

region_source

source of the region ('EEZ','MPA', 'RFMO' or 'USER_JSON')

key

Authorization token. Can be obtained with gfw_auth() function

Details

The user-defined geojson has to be surrounded by a geojson tag, that can be created using a simple paste:

geojson_tagged <- paste0('{"geojson":', your_geojson,'}').

If you have an sf shapefile, you can also use function sf_to_geojson() to obtain the correctly-formatted geojson.

Examples

library(gfwr)
code_eez <- get_region_id(region_name = 'CIV', region_source = 'EEZ',
key = gfw_auth())
get_raster(spatial_resolution = 'LOW',
           temporal_resolution = 'YEARLY',
           group_by = 'FLAG',
           date_range = '2021-01-01,2021-10-01',
           region = code_eez$id,
           region_source = 'EEZ',
           key = gfw_auth())
#> Rows: 611 Columns: 6
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): flag
#> dbl (5): Lat, Lon, Time Range, Vessel IDs, Apparent Fishing Hours
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 611 × 6
#>      Lat   Lon `Time Range` flag  `Vessel IDs` `Apparent Fishing Hours`
#>    <dbl> <dbl>        <dbl> <chr>        <dbl>                    <dbl>
#>  1   4.9  -5.7         2021 CHN              4                    37.3 
#>  2   3.2  -5.8         2021 BLZ              1                     2.03
#>  3   2.2  -6.7         2021 BLZ              1                     0.96
#>  4   1.9  -6.9         2021 BLZ              1                     1.19
#>  5   1.2  -6.8         2021 BLZ              1                     2.46
#>  6   5    -5.4         2021 CHN              3                    23.8 
#>  7   4.6  -4.7         2021 SLV              1                     2.38
#>  8   5.3  -4           2021 BES              4                    67.0 
#>  9   5.3  -4           2021 SLV              3                    17.9 
#> 10   5.1  -4.2         2021 BLZ              1                     1.99
#> # ℹ 601 more rows